+++ server/data-service/src/main/java/edu/illinois/ncsa/incore/service/data/controllers/FileController.java
@@ -10,7 +10,6 @@
 
 package edu.illinois.ncsa.incore.service.data.controllers;
 
-import edu.illinois.ncsa.incore.common.config.Config;
 import edu.illinois.ncsa.incore.common.exceptions.IncoreHTTPException;
 import edu.illinois.ncsa.incore.service.data.dao.IRepository;
 import edu.illinois.ncsa.incore.service.data.models.Dataset;
@@ -33,7 +32,7 @@ import java.util.List;
 @Path("files")
 public class FileController {
     private Logger logger = Logger.getLogger(edu.illinois.ncsa.incore.service.data.controllers.FileController.class);
-    private static final String DATA_REPO_FOLDER = Config.getConfigProperties().getProperty("data.repo.data.dir");
+    private static final String DATA_REPO_FOLDER = System.getenv("DATA_REPO_DATA_DIR");
 
     @Inject
     private IRepository repository;