+++ server/dfr3-service/src/main/java/edu/illinois/ncsa/incore/service/dfr3/CorsFilter.java
@@ -21,7 +21,12 @@ public class CorsFilter implements ContainerResponseFilter {
     public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext) throws IOException {
         MultivaluedMap<String, Object> headers = responseContext.getHeaders();
 
-//        headers.add("Access-Control-Allow-Origin", "*");
+        // if header origin is in the allowed origin list; add it to access-control-allow-origin header
+        String allowedOrigins = System.getenv("SERVICES_ALLOW_ORIGIN");
+        String requestHeadersOrigin = requestContext.getHeaderString("Origin");
+        if (requestHeadersOrigin != null && allowedOrigins.contains(requestHeadersOrigin)){
+            headers.add("Access-Control-Allow-Origin", requestHeadersOrigin);
+        }
         headers.add("Access-Control-Allow-Methods", "GET");
         headers.add("Access-Control-Allow-Headers", "Access-Control-Allow-Headers, Origin, Accept, X-Requested-With, " +
             "Content-Type, Access-Control-Request-Method, Access-Control-Request-Headers, auth-user, " +